--- bin/manual_config_pdf
@@ -1,16 +1,0 @@
-#!/usr/bin/env bash
-
-# if there is <only one> manual in the available manuals array below,
-# the manual directory is set to ROOT and not the manual name dynamically generated from the array list
-# also, set in that case the postfix_name to "" and have the manual name like you want to print it (e.g. "Desktop_Client")
-# only change ROOT in case you know what you are doing
-# be aware that a manual name MUST NOT contain a blank
-MANUAL_ROOT_NAME="ROOT"
-# the postfix component is used to create the path to the manual like admin_manual 
-MANUAL_POSTFIX_NAME="_manual"
-# note that this name must be part of the book filename in books/
-# a part of the file name is composed by available_manuals+manual_postfix
-AVAILABLE_MANUALS=(admin user developer)
-# define if the first letter of the manual name must be captitalized (Desktop)
-# make empty in case it can be also lower case (like iOS-App)
-FIRST_LETTER_CAPITAL="Y"
