+++ site.yml
@@ -9,6 +9,9 @@ content:
   - url: .
     branches:
     - HEAD
+  # note that there are no docs-xxx sources present for the test build to satisfy used references
+  # we therefore have set failure_level: none, see below for more info
+  # the real build is made in the docs repo using correct values
 
 ui:
   output_dir: assets
@@ -31,9 +34,6 @@ asciidoc:
     experimental: true
     # Antora pagination (prev page, next page)
     page-pagination: true
-    # Custom attribute page-component-build-list
-    # Only embed named opengraph images in docs-ui via src/partials/head-meta-opengraph.hbs sourced from docs/overlay
-    page-component-build-list: 'android, branding, desktop, docs, ios-app, ocis, server, user, webui'
 #   common
     docs-base-url: 'https://doc.owncloud.com'
     oc-complete-base-url: 'https://download.owncloud.com/server/stable'
@@ -42,14 +42,18 @@ asciidoc:
     oc-marketplace-url: 'https://marketplace.owncloud.com'
     oc-central-url: 'https://central.owncloud.org'
     oc-support-url: 'https://owncloud.com/support'
+# note that the version attributes just need to be present and have 'next' as value
+# the same applies for other repo dependent attributes as they are just here for
+# local building, only change them if there are local building issues!
+#
 #   docs
     latest-docs-version: 'next'
     previous-docs-version: 'next'
 #   server
-    latest-server-version: '10.13'
-    latest-server-download-version: '10.13.4'
-    previous-server-version: '10.12'
-    current-server-version: '10.13'
+    latest-server-version: 'next'
+    previous-server-version: 'next'
+    latest-server-download-version: 'next'
+    current-server-version: 'next'
     oc-changelog-url: 'https://owncloud.com/changelog/server/'
     oc-install-package-url: 'https://download.owncloud.com/server/stable/?sort=time&order=asc'
     oc-examples-server-url: 'https://owncloud.install.com/owncloud'
@@ -78,19 +82,32 @@ asciidoc:
     latest-webui-version: 'next'
     previous-webui-version: 'next'
 #   desktop
-    latest-desktop-version: '5.2'
-    previous-desktop-version: '5.1'
+    latest-desktop-version: 'next'
+    previous-desktop-version: 'next'
 #   ios-app
-    latest-ios-version: '12.0'
-    previous-ios-version: '11.11'
+    latest-ios-version: 'next'
+    previous-ios-version: 'next'
 #   android
-    latest-android-version: '4.1'
-    previous-android-version: '4.0'
+    latest-android-version: 'next'
+    previous-android-version: 'next'
 #   branded
     latest-branded-version: 'next'
     previous-branded-version: 'next'
   extensions:
-    - ./lib/extensions/tabs.js
-    - ./lib/extensions/remote-include-processor.js
-    - ./node_modules/asciidoctor-kroki/src/asciidoctor-kroki.js
+    - ./ext-asciidoc/tabs.js
+    - ./ext-asciidoc/remote-include-processor.js
+    - asciidoctor-kroki
 
+antora:
+  extensions:
+    - ./ext-antora/generate-index.js
+
+runtime:
+  log:
+    # https://docs.antora.org/antora/latest/playbook/runtime-log-failure-level/#failure-level-key
+    # e.g. xrefs will report errors but building will now return 0 (ok)
+    # using this prevents CI from never getting green. in antora 2, xrefs errors always returned 0
+    # this is not an ideal solution and should be removed when the partial build extension (Atlas) is available
+    # see: https://gitlab.com/antora/antora-atlas-extension
+    # drawback, we have to (regulary) monitor the docs repo for build errors origin in docs-main via a local build.
+    failure_level: none
